+++ source/computer/sequentialComputer.cpp
@@ -24,7 +24,7 @@ void SequentialComputer::fillForces()
 }
 
 
-void SequentialComputer::init(std::vector<Particle> &particles, size_t _N)
+void SequentialComputer::init(std::vector<Particle> &particles, ull _N)
 {
 	particleVectors[previous] = particles;
 	particleVectors[current] = particles;
